+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 2018/04/13
 
     - use `ipairs` instead of `pairs` to traverse lists of images and image match functions
+    - load extensions in the correct order
 
 - 2018/04/09